Updates per PR comments
