+++ utils/shared.js
@@ -792,6 +792,219 @@ const cleanSurveyData = (data) => {
     return data;
 }
 
+/**
+ * Gets baseline data updates for participants when submitting specimens
+ * 
+ * @param {object} biospecimenData The biospecimen data
+ * @param {object} participantData The participant data
+ * @param {array} specimenArray The array of specimens for the participant
+ * @param {array} siteTubesList The array of tubes used for the site
+ * @returns {object}
+ */
+const updateBaselineData = (biospecimenData, participantData, participantUid, specimenArray, siteTubesList) => {
+    let participantUpdates = {};
+    let settings = {};
+    let visit = biospecimenData[fieldMapping.collectionSelectedVisit];
+    // Now we potentially need to updateBaselineData
+    const baselineVisit = (biospecimenData[fieldMapping.collectionSelectedVisit] === fieldMapping.baseline);
+    const clinicalResearchSetting = (biospecimenData[fieldMapping.collectionSetting] === fieldMapping.research || biospecimenData[fieldMapping.collectionSetting] === fieldMapping.clinical);
+    if (baselineVisit && clinicalResearchSetting) {
+        // Update baseline data
+        const baselineCollections = specimenArray.filter(specimen => specimen[fieldMapping.collectionSelectedVisit] === fieldMapping.baseline);
+
+        const bloodTubes = siteTubesList.filter(tube => tube.tubeType === "Blood tube");
+        const urineTubes = siteTubesList.filter(tube => tube.tubeType === "Urine");
+        const mouthwashTubes = siteTubesList.filter(tube => tube.tubeType === "Mouthwash");
+
+        let bloodCollected = (participantData[fieldMapping.baselineBloodSampleCollected] === fieldMapping.yes);
+        let urineCollected = (participantData[fieldMapping.baselineUrineCollected] === fieldMapping.yes);
+        let mouthwashCollected = (participantData[fieldMapping.baselineMouthwashCollected] === fieldMapping.yes);
+        let allBaselineCollected = (participantData[fieldMapping.allBaselineSamplesCollected] === fieldMapping.yes);
+
+        let bloodTubesLength = 0
+        let urineTubesLength = 0
+        let mouthwashTubesLength = 0
+
+        const collectionSetting = biospecimenData[fieldMapping.collectionSetting];
+        const isResearch = collectionSetting === fieldMapping.research;
+        const isClinical = collectionSetting === fieldMapping.clinical;
+
+        // Build the collection details
+        if (participantData[fieldMapping.collectionDetails]) {
+            settings = participantData[fieldMapping.collectionDetails];
+            if (!settings[visit]) settings[visit] = {};
+    
+        } else {
+            settings = {
+                [visit]: {}
+            }
+        }
+    
+        if (!settings[visit][fieldMapping.bloodCollectionSetting]) {
+            bloodTubes.forEach(tube => {
+                const tubeIsCollected = biospecimenData[tube.concept][fieldMapping.tubeIsCollected] === fieldMapping.yes;
+                if(tubeIsCollected) {
+                    settings[visit][fieldMapping.bloodCollectionSetting] = collectionSetting;
+                    if(isResearch) {
+                        settings[visit][fieldMapping.baselineBloodCollectedTime] = biospecimenData[fieldMapping.collectionDateTimeStamp];
+                    }
+                    else if(isClinical) {
+                        settings[visit][fieldMapping.clinicalBloodCollected] = fieldMapping.yes;
+                        settings[visit][fieldMapping.clinicalBloodCollectedTime] = biospecimenData[fieldMapping.collectionScannedTime];
+    
+                        settings[visit][fieldMapping.anySpecimenCollected] = fieldMapping.yes;
+    
+                        if(!(settings[visit][fieldMapping.anySpecimenCollectedTime])) {
+                            settings[visit][fieldMapping.anySpecimenCollectedTime] = biospecimenData[fieldMapping.collectionScannedTime];
+                        }
+                    }
+                    bloodTubesLength += 1
+                }
+            });
+        }
+        else if (settings[visit][fieldMapping.baselineBloodCollectedTime] !== '' ||  settings[visit][fieldMapping.clinicalBloodCollectedTime] !== ''){
+            const participantBloodCollected = participantData[fieldMapping.baselineBloodSampleCollected] === fieldMapping.yes;
+            const totalBloodTubesAvail = bloodTubes.filter((tube) => biospecimenData[tube.concept][fieldMapping.tubeIsCollected] === fieldMapping.yes);
+            if (totalBloodTubesAvail.length === 0 && participantBloodCollected) {
+                delete settings[visit][fieldMapping.bloodCollectionSetting]; // derived variables & timestamp are updated only if all the blood tubes are unchecked
+                if (isResearch) {
+                    delete settings[visit][fieldMapping.baselineBloodCollectedTime];
+                }
+                else if (isClinical) {
+                    settings[visit][fieldMapping.clinicalBloodCollected] = fieldMapping.no;
+                    delete settings[visit][fieldMapping.clinicalBloodCollectedTime];
+    
+                    if (urineTubesLength === 0 && mouthwashTubesLength === 0) { // anySpecimenCollected variable will only be updated to NO if mouthwash & urine specimens are not present.
+                        settings[visit][fieldMapping.anySpecimenCollected] = fieldMapping.no;
+                        if (!(settings[visit][fieldMapping.anySpecimenCollectedTime])) {
+                            delete settings[visit][fieldMapping.anySpecimenCollectedTime];
+                        }
+                    }
+                }
+                participantUpdates[fieldMapping.baselineBloodSampleCollected] = fieldMapping.no;
+                bloodTubesLength = totalBloodTubesAvail.length;
+            }
+        }
+    
+        if (!settings[visit][fieldMapping.urineCollectionSetting]) {
+            urineTubes.forEach(tube => {
+                const tubeIsCollected = biospecimenData[tube.concept][fieldMapping.tubeIsCollected] === fieldMapping.yes;
+                if (tubeIsCollected) {
+                    settings[visit][fieldMapping.urineCollectionSetting] = collectionSetting;
+                    if (isResearch) {
+                        settings[visit][fieldMapping.baselineUrineCollectedTime] = biospecimenData[fieldMapping.collectionDateTimeStamp];
+                    }
+                    else if (isClinical) {
+                        settings[visit][fieldMapping.clinicalUrineCollected] = fieldMapping.yes;
+                        settings[visit][fieldMapping.clinicalUrineCollectedTime] = biospecimenData[fieldMapping.collectionScannedTime];
+    
+                        settings[visit][fieldMapping.anySpecimenCollected] = fieldMapping.yes;
+    
+                        if (!(settings[visit][fieldMapping.anySpecimenCollectedTime])) {
+                            settings[visit][fieldMapping.anySpecimenCollectedTime] = biospecimenData[fieldMapping.collectionScannedTime];
+                        }
+                    }
+                    urineTubesLength += 1
+                }
+            });
+        }
+        else if (settings[visit][fieldMapping.baselineUrineCollectedTime] !== '' ||  settings[visit][fieldMapping.clinicalUrineCollectedTime] !== '') {
+            const participantUrineCollected = participantData[fieldMapping.baselineUrineCollected] === fieldMapping.yes;
+            const totalUrineTubesAvail = urineTubes.filter((tube) => biospecimenData[tube.concept][fieldMapping.tubeIsCollected] === fieldMapping.yes);
+            if (totalUrineTubesAvail.length === 0 && participantUrineCollected) {
+                delete settings[visit][fieldMapping.urineCollectionSetting];
+                if(isResearch) {
+                    delete settings[visit][fieldMapping.baselineUrineCollectedTime];
+                }
+                else if (isClinical) {
+                    settings[visit][fieldMapping.clinicalUrineCollected] = fieldMapping.no;
+                    delete settings[visit][fieldMapping.clinicalUrineCollectedTime];
+    
+                    if (bloodTubesLength === 0 && mouthwashTubesLength === 0) { // anySpecimenCollected variable will only be updated to NO if mouthwash & blood specimens are not present.
+                        settings[visit][fieldMapping.anySpecimenCollected] = fieldMapping.no;
+                        if (!(settings[visit][fieldMapping.anySpecimenCollectedTime])) {
+                            delete settings[visit][fieldMapping.anySpecimenCollectedTime];
+                        }
+                    }
+                }
+                derivedVariables[fieldMapping.baselineUrineCollected] = fieldMapping.no;
+                urineTubesLength = totalUrineTubesAvail.length;
+            }  
+        }
+    
+        if (!settings[visit][fieldMapping.mouthwashCollectionSetting]) {
+            mouthwashTubes.forEach(tube => {
+                const isTubeCollected = biospecimenData[tube.concept][fieldMapping.tubeIsCollected] === fieldMapping.yes;
+                if (isTubeCollected) {
+                    settings[visit][fieldMapping.mouthwashCollectionSetting] = collectionSetting;
+                    if (isResearch) {
+                        settings[visit][fieldMapping.baselineMouthwashCollectedTime] = biospecimenData[fieldMapping.collectionDateTimeStamp];
+                    }
+                mouthwashTubesLength += 1
+                }
+            });
+        }
+        else if (settings[visit][fieldMapping.baselineMouthwashCollectedTime] !== '' && participantData[fieldMapping.baselineMouthwashCollected] === fieldMapping.yes) {
+            const isParticipantMouthwashCollected = participantData[fieldMapping.baselineMouthwashCollected] === fieldMapping.yes;
+            const totalMouthwasTubesAvail = mouthwashTubes.filter((tube) => biospecimenData[tube.concept][fieldMapping.tubeIsCollected] === fieldMapping.yes);
+            if (totalMouthwasTubesAvail.length === 0 &&  isParticipantMouthwashCollected) {
+                delete settings[visit][fieldMapping.mouthwashCollectionSetting]
+                if (isResearch) {
+                    delete settings[visit][fieldMapping.baselineMouthwashCollectedTime];
+                }
+                derivedVariables[fieldMapping.baselineMouthwashCollected] = fieldMapping.no;
+                mouthwashTubesLength = totalMouthwasTubesAvail.length;
+            }
+        }
+
+        participantUpdates[fieldMapping.collectionDetails] = settings;
+
+        baselineCollections.forEach(collection => {
+
+            if (!bloodCollected) {
+                bloodTubes.forEach(tube => {
+                    if (collection[tube.concept]?.[fieldMapping.tubeIsCollected] === 353358909) {
+                        bloodCollected = true;
+                    }
+                });
+            } 
+            if (!urineCollected) {
+                urineTubes.forEach(tube => {
+                    if (collection[tube.concept]?.[fieldMapping.tubeIsCollected] === 353358909) {
+                        urineCollected = true;
+                    }
+                });
+            }
+            if (!mouthwashCollected) {
+                mouthwashTubes.forEach(tube => {
+                    if (collection[tube.concept]?.[fieldMapping.tubeIsCollected] === 353358909) {
+                        mouthwashCollected = true;
+                    }
+                });
+            }
+            
+        });
+
+        if (baselineCollections.length > 0 && baselineCollections[0][fieldMapping.collectionSetting] === fieldMapping.research) {
+            allBaselineCollected = bloodCollected && urineCollected && mouthwashCollected;
+        }
+        else if (baselineCollections.length > 0 && baselineCollections[0][fieldMapping.collectionSetting] === fieldMapping.clinical) {
+            allBaselineCollected = bloodCollected && urineCollected;
+        }
+
+        participantUpdates = {
+            ...participantUpdates,
+            [fieldMapping.baselineBloodSampleCollected]: bloodCollected ? fieldMapping.yes : fieldMapping.no,
+            [fieldMapping.baselineUrineCollected]: urineCollected ? fieldMapping.yes : fieldMapping.no,
+            [fieldMapping.baselineMouthwashCollected]: mouthwashCollected ? fieldMapping.yes : fieldMapping.no,
+            [fieldMapping.allBaselineSamplesCollected]: allBaselineCollected ? fieldMapping.yes : fieldMapping.no,
+            uid: participantUid
+        };
+
+    }
+    return participantUpdates
+}
+
 const convertSiteLoginToNumber = (siteLogin) => {
     const siteLoginNumber = parseInt(siteLogin);
     if (siteLoginNumber === NaN) return undefined;
@@ -1692,6 +1905,7 @@ module.exports = {
     sites, 
     bagConceptIDs,
     cleanSurveyData,
+    updateBaselineData,
     refusalWithdrawalConcepts,
     convertSiteLoginToNumber,
     swapObjKeysAndValues,
